+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - Fixed web minidump upload to work with zero attachments
+- Fixed web crash list from showing all rows as "Processing"
 
 
 ## [0.3.0] - 2021-09-29