+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.1] - 2021-10-29
+### Removed
+- `charset-normalizer`: Switched to using BytesIO
+
 ### Fixed
 - Fixed web minidump upload to work with zero attachments
 - Fixed web crash list from showing all rows as "Processing"